+++ src/main/java/by/laba1/demo/core/service/PatientService.java
@@ -2,17 +2,19 @@ package by.laba1.demo.core.service;
 
 import by.laba1.demo.api.dto.patient.CreatePatientDto;
 import by.laba1.demo.api.dto.patient.GetPatientDto;
+import by.laba1.demo.api.error.MessageException;
+import by.laba1.demo.api.error.ResourceNotFoundException;
 import by.laba1.demo.core.dao.chmem.MyCache;
 import by.laba1.demo.core.dao.patient.PatientRepository;
 import by.laba1.demo.core.entities.Patient;
 import by.laba1.demo.core.mapper.patient.CreatePatientMapper;
 import by.laba1.demo.core.mapper.patient.GetPatientMapper;
+import jakarta.persistence.EntityNotFoundException;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.HttpStatus;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.stereotype.Service;
-import org.springframework.web.server.ResponseStatusException;
 
 @Service
 @RequiredArgsConstructor(onConstructor_ = {@Autowired})
@@ -20,13 +22,13 @@ public class PatientService {
     private final PatientRepository patientRepository;
     private final CreatePatientMapper createPatientMapper;
     private final GetPatientMapper getPatientMapper;
-    private final MyCache<Long, GetPatientDto> patientMyCache = new MyCache<>(10 * 60 * 1000L);
+    private final MyCache<Long, GetPatientDto> patientMyCache = new MyCache<>(10, 120_000);
 
     public List<GetPatientDto> getPatientsByFilter(String name, String phoneNumber) {
         List<Patient> patients = patientRepository.findByFilters(name, phoneNumber);
 
         if (patients.isEmpty()) {
-            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Patients not found");
+            throw new EntityNotFoundException(MessageException.ENTITY_WITH_CRITERIA_NOT_FOUND);
         }
 
         return getPatientMapper.toDtos(patients);
@@ -39,8 +41,9 @@ public class PatientService {
         }
 
         Patient patientFound = patientRepository.findById(id).orElseThrow(
-            () -> new ResponseStatusException(HttpStatus.NOT_FOUND, "Patients not found")
-        );
+            () -> new EntityNotFoundException(
+                String.format(MessageException.ENTITY_WITH_ID_NOT_FOUND, id)
+        ));
         GetPatientDto dto = getPatientMapper.toDto(patientFound);
         patientMyCache.put(id, dto);
 
@@ -49,8 +52,7 @@ public class PatientService {
 
     public GetPatientDto createPatient(CreatePatientDto dto) {
         if (patientRepository.existsByPhoneNumber(dto.getPhoneNumber())) {
-            throw new ResponseStatusException(HttpStatus.CONFLICT,
-                "Patient with this email already exist: " + dto.getPhoneNumber());
+            throw new DataIntegrityViolationException(MessageException.PHONE_NUMBER_UNIQUE + dto.getPhoneNumber());
         }
 
         Patient patient = createPatientMapper.toEntity(dto);
@@ -60,13 +62,11 @@ public class PatientService {
 
     public GetPatientDto updatePatient(long id, CreatePatientDto dto) {
         Patient patient = patientRepository.findById(id)
-            .orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND,
-                "Patient with this id " + id + " not found"));
+            .orElseThrow(() -> new EntityNotFoundException("Patient not found with ID " + id));
 
         if (!patient.getPhoneNumber().equals(dto.getPhoneNumber()) &&
             patientRepository.existsByPhoneNumber(dto.getPhoneNumber())) {
-            throw new ResponseStatusException(HttpStatus.CONFLICT,
-                "Patient with this phone number already exist: " + dto.getPhoneNumber());
+            throw new DataIntegrityViolationException(MessageException.PHONE_NUMBER_UNIQUE + dto.getPhoneNumber());
         }
         
         createPatientMapper.merge(patient, dto);
@@ -76,6 +76,9 @@ public class PatientService {
     }
 
     public void deletePatient(long id) {
+        if (!patientRepository.existsById(id)) {
+            throw new ResourceNotFoundException(String.format(MessageException.ENTITY_WITH_ID_NOT_FOUND, id));
+        }
         patientRepository.deleteById(id);
         patientMyCache.clear();
     }
